% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{Dnichenetwork}
\alias{Dnichenetwork}
\title{Calculate niche-niche connections with connectome data from nichetag method}
\usage{
Dnichenetwork(
  scObject,
  groupby = "cell_clusters",
  share_method = "min",
  highconfidence = TRUE,
  send_cutoff1 = 10,
  send_cutoff2 = 0.8,
  direction = FALSE
)
}
\arguments{
\item{scObject}{Seurat, input seurat object in which tag expression matrix is merged in gene expression matrix}

\item{groupby}{character, a colname in meta data of seurat object, used for set defination}

\item{share_method}{string, min,max,mean, the method to calculate connection strength between different sets}

\item{highconfidence}{TorF, if or not to distinguish high confident tag-sender}

\item{send_cutoff1}{int, count cutoff for primary celltag asignment}

\item{send_cutoff2}{numeric, percentage cutoff for primary celltag asignment}

\item{direction}{TorF, if or not to distinguish tag-sender or tag-receiver}
}
\value{
a list contains connectome information of all niches:
\describe{
  \item{tag_matrix}{data.frame, tag expression matrix of all cells}
  \item{cell_type}{factor, cell type or state used for set defination}
  \item{cell_tags}{numeric, total expression of each tag}
  \item{niche_size}{integer, covered cell number of each tag}
  \item{niche_celltypes}{integer, covered cell types or cell states of each tag}
  \item{niche_interaction}{data.frame, overlapped cell numbers of tag-tag pairs}
  \item{code2celltype}{character,corresponding celltype of each celltypes code positons}
  \item{code2tag}{character, corresponding tag of each tag code positons}
  \item{cell_barcode}{character, code of each cell, tagcode+celltypecode}
  \item{set}{list, each element represent a tag expression matrix of the set}
  \item{set_size}{integer, covered cell number of each set}
  \item{niche}{list, mean tag expression of covered cells of all covered set of each tag}
  \item{set_interaction}{data.frame, overlapped tag expression of set-set pairs}
  \item{setID}{integer, order of sets when make graph with graph_from_data_frame in igraph }
}
}
\description{
This fuction use tag expression matrix to infer the connection among niches labeled with different nichetags
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
summary(nnt)
nnt=Dnichenetwork(scObject,groupby="cell_clusters",share_method="mean")
dnnt=Dnichenetwork(scObject,groupby="cell_clusters",direction = TRUE)
}
