% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{Gephi_prepare}
\alias{Gephi_prepare}
\title{Extract set connections from nnt for Gephi analysis}
\usage{
Gephi_prepare(nnt, file = "Gephi_input.xlsx")
}
\arguments{
\item{nnt}{a list contains connectome information of all niches}

\item{file}{output file name}
}
\value{
a file for Gephi to get more circular layout
}
\description{
Extract set connections from nnt for Gephi analysis
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
Gephi_prepare(nnt)
}
