% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{print_clustertag_v1}
\alias{print_clustertag_v1}
\title{print cluster vs tag number and types}
\usage{
print_clustertag_v1(nnt, file = "clustertag.pdf")
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf}
}
\value{
pdf
}
\description{
print cluster vs tag number and types
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
print_clustertag(nnt)
}
