% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{set_expr}
\alias{set_expr}
\title{Get set expression matrix, for each gene use mean value of all cells of the same set}
\usage{
set_expr(scObject, nnt, seurat_layer = "counts")
}
\arguments{
\item{scObject}{Seurat, input seurat object in which tag expression matrix is merged in gene expression matrix}

\item{nnt}{a list contains connectome information of all niches}

\item{seurat_layer}{layer in scObject used to calculate expression level of genes}
}
\value{
set expression matrix of all nontag genes
}
\description{
Get set expression matrix, for each gene use mean value of all cells of the same set
}
\examples{
data(example)
nnt <- Dnichenetwork(scObject, groupby = "cell_clusters")
setmatrix=set_expr(scObject,nnt,seurat_layer="counts")
}
