% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{settype}
\alias{settype}
\title{print set type distribution for each niche}
\usage{
settype(nnt, file = "settype.pdf", onlyHighconfident = TRUE)
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf}

\item{onlyHighconfident}{TorF, if or not use remove sets of celltag indeterminate cancer cells}
}
\value{
pdf
}
\description{
print set type distribution for each niche
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
settype(nnt)
}
