% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{settype_v1}
\alias{settype_v1}
\title{print set type distribution for each niche}
\usage{
settype_v1(nnt, file = "settype.pdf")
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf}
}
\value{
pdf
}
\description{
print set type distribution for each niche
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
settype(nnt)
}
