% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{tag_cci}
\alias{tag_cci}
\title{print sender and receiver cell numbers for each tag, cutoff for sender is 10 and 0.8}
\usage{
tag_cci(nnt, send_cutoff1 = 10, send_cutoff2 = 0.8, file = "tag2ccitypes.pdf")
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{send_cutoff1}{int, count cutoff for primary celltag asignment}

\item{send_cutoff2}{numeric, percentage cutoff for primary celltag asignment}

\item{file}{pdf}
}
\value{
pdf
}
\description{
print sender and receiver cell numbers for each tag, cutoff for sender is 10 and 0.8
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
tag_cci(nnt)
}
