% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{tag_cci_v1}
\alias{tag_cci_v1}
\title{print sender and receiver cell numbers for each tag, cutoff is 3}
\usage{
tag_cci_v1(nnt, file = "tag2ccitypes.pdf")
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf}
}
\value{
pdf
}
\description{
print sender and receiver cell numbers for each tag, cutoff is 3
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
tag_cci(nnt)
}
