% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{tag_cellsettype}
\alias{tag_cellsettype}
\title{print cell type and set type distribution for each niche}
\usage{
tag_cellsettype(nnt, file = "tag2celltype_settype.pdf", seed = 888)
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf}

\item{seed}{seed for color palette creation}
}
\value{
pdf
}
\description{
print cell type and set type distribution for each niche
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
tag_cellsettype(nnt)
}
